Provides the ability to easily queue deletions
In the effort to make the task queueing code more generic, it was
forgotten that an important aspect of the previous code was
special logic for deletes that did not instantiate the object
from the database.  In order to maintain the ability to keep
the code generic, an argument was added to submission that
allows the caller to define how an instance is instantiated on
the worker.  In the case of a delete, we know all we'll need
is the pk, so we make a simple object witht he pk as an
attribute, and return that.